Found the source of vgg16 not learning, was the choose_model version. there are currently three versions in the modelManagement py file. The GOOD version is: choose_model1. As part of this investigation, functions have been moved into appropriate py files and py files created as necessary so that no file is too large. Names are hopefully intuitive except for functions which should be imageProcessingFunctions
